# optimize regex handling