I think these are good changes, but the perf results are exactly the same as my non-perf-related PR:

```
       django_template.py             3.6s (2)             3.6s (2)  -0.2%
            pyxl_bench.py             3.5s (2)             3.4s (2)  -1.2%
sqlalchemy_imperative2.py             4.3s (2)             4.4s (2)  +0.4%
                  geomean                 3.8s                 3.8s  -0.4%
```

It speeds up some simple regex microbenchmarks by 4x though.