@kmod 2015-07-17
It looks like Travis-CI tested the wrong commit, but I ran the tests manually and it seems ok.

Here are the stats on the benchmarks I was looking at:
```
         django_lexing.py             2.0s (2)             1.8s (2)  -11.6%
      re_split_ubench2.py             2.9s (2)             1.8s (2)  -38.7%
       re_split_ubench.py             3.1s (2)             0.8s (2)  -75.3%
```
django_lexing is the lexing part of the template parsing; re_split_ubench2 is the regex portion of the lexing, and re_split_ubench is just a tiny regex (something like `re.split("banana" * 100, "ana")`).